requirements.txt
flask

# YOLOv5 requirements
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
seaborn>=0.11.0
pandas

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop>=0.1.1  # FLOPs computation

argparse==1.4.0
Flask==1.1.2
numpy==1.20.2
tensorflow==2.5.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
Keras-Applications==1.0.8
urllib3==1.26.4
Werkzeug==1.0.1
